requirements.txt
anyio==4.0.0
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.1.3
firebase-admin==6.2.0
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jws==0.1.3
msgpack==1.0.5
oauth2client==3.0.0
packaging==23.1
pipdeptree==2.13.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.8.0
pyparsing==3.1.1
python-jwt==2.0.1
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.5.0
openai==1.3.5
